Move ProcessState to job module
We are going to use this type in the System trait, so it should be
defined outside of the virtual system module.
