+++ yash-semantics/src/command/pipeline.rs
@@ -335,12 +335,12 @@ mod tests {
     use std::rc::Rc;
     use yash_env::builtin::Builtin;
     use yash_env::builtin::Type::Special;
+    use yash_env::job::ProcessState;
     use yash_env::job::WaitStatus;
     use yash_env::option::Option::Monitor;
     use yash_env::option::State::On;
     use yash_env::semantics::Field;
     use yash_env::system::r#virtual::FileBody;
-    use yash_env::system::r#virtual::ProcessState;
     use yash_env::trap::Signal;
     use yash_env::VirtualSystem;
